+++ lib/picotls.c
@@ -30,7 +30,6 @@
 #include <sys/time.h>
 #endif
 #include "picotls.h"
-#include "picotls/pembase64.h"
 
 #define PTLS_MAX_PLAINTEXT_RECORD_SIZE 16384
 #define PTLS_MAX_ENCRYPTED_RECORD_SIZE (16384 + 256)
@@ -3760,20 +3759,3 @@ static uint64_t get_time(ptls_get_time_t *self)
 
 ptls_get_time_t ptls_get_time = {get_time};
 
-#define PTLS_MAX_CERTS_IN_CONTEXT 16
-
-int ptls_load_certificates(ptls_context_t *ctx, char *cert_pem_file)
-{
-    int ret = 0;
-
-    ctx->certificates.list = (ptls_iovec_t *)malloc(PTLS_MAX_CERTS_IN_CONTEXT * sizeof(ptls_iovec_t));
-
-    if (ctx->certificates.list == NULL) {
-        ret = PTLS_ERROR_NO_MEMORY;
-    } else {
-        ret = ptls_load_pem_objects(cert_pem_file, "CERTIFICATE", ctx->certificates.list, PTLS_MAX_CERTS_IN_CONTEXT,
-                                    &ctx->certificates.count);
-    }
-
-    return ret;
-}
